# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: string;
  name: string;
  client: string;
  location: string;
  completion: number;
  updatedAt: string;
}

export interface SpecificationInput {
  environment: string;
  item: string;
  dimension: string;
  finish: string;
  brand: string;
  budget: number;
  quotedPrice: number;
}

export type Specification = SpecificationInput & {
  id: string;
  updatedAt: string;
};

export type ProjectDetail = Project & {
  specifications: Specification[];
};