# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: string;
  name: string;
  client: string;
  location: string;
  completion: number;
  updatedAt: string;
}