# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Project } from './project';
import type { Specification } from './specification';

export type ProjectDetail = Project & {
  specifications: Specification[];
};