# lib/api-zod/src/generated/types/specification.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SpecificationInput } from './specificationInput';

export type Specification = SpecificationInput & {
  id: string;
  updatedAt: string;
};