# lib/api-zod/src/generated/types/specificationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SpecificationInput {
  environment: string;
  item: string;
  dimension: string;
  finish: string;
  brand: string;
  budget: number;
  quotedPrice: number;
}